requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
josepy==1.13.0
mozilla-django-oidc==3.0.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
